scripts/requirements.txt
python-frontmatter>=1.1.0
markdown>=3.5
python-dotenv>=1.0
supabase>=2.0
